# docs/openapi.yaml
openapi: 3.1.0
info:
  title: discord-support-hub API
  version: 0.1.0
  summary: API-driven provisioning and governance of isolated private Discord support spaces.
  description: |
    The control surface (FR-11) for discord-support-hub, consumed by the backoffice
    (no human UI in v1). Postgres is the authorization source of truth; Discord is a
    projection the bot reconciles (three-layer truth model).

    Conventions:
      * **Mutating endpoints are asynchronous.** They validate, persist the *desired*
        state, enqueue a job, and return `202 Accepted` with a `job` handle. Poll
        `GET /jobs/{jobId}` for completion. (NFR-2: latency decoupled from Discord rate limits.)
      * **Read endpoints** are served from cache (TTL + write-invalidation) and return `200`.
      * **Authorization always resolves against Postgres**, never the Discord role (NFR-13).
      * **Idempotency:** mutating requests accept an `Idempotency-Key` header; a repeated key
        replays the original response instead of re-enqueueing (NFR-3).
      * **Role-per-merchant access.** Provisioning a space auto-creates a Discord role for the
        merchant and grants that role VIEW+SEND on the channel. A collaborator acquires access by
        joining through the merchant's native invite-with-role link (created once by the operator
        in the Discord client and stored via `PUT /merchants/{merchantId}/invite`), which the hub
...
  license:
    name: Apache-2.0
    identifier: Apache-2.0
...
servers:
...
security:
  - ServiceApiKey: []
...
paths:
...
  /channels/{id}/collaborators:
    post:
      tags: [Collaborators]
      operationId: registerCollaborator
      summary: Register a collaborator on a space by name + work email (synchronous).
      description: |
        Records the desired `space_member` plus the collaborator's **name and work email** (our
        traceability labels, stored in Postgres — never a Discord lookup key). This is a synchronous
        roster write with **no Discord side-effect**: access is role-based and is acquired later when
        the collaborator joins through the merchant's invite-with-role link. Returns `201`.

        Send the invite with `POST /channels/{id}/collaborators/{userId}:send-invite`. Only
        Agents/Admins may register collaborators (FR-20).
      parameters:
        - $ref: '#/components/parameters/SpaceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterCollaboratorRequest'
      responses:
        '201':
          description: Collaborator registered on the space (invite not yet sent).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpaceMember' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /channels/{id}/collaborators/{userId}:send-invite:
    post:
      tags: [Collaborators]
      operationId: sendCollaboratorInvite
      summary: Email the merchant's invite-with-role link to a registered collaborator (async).
      description: |
        Enqueues a `notify` job that emails the merchant's stored invite-with-role link (with the
        configurable welcome message) to the collaborator's work email via the hub's SMTP relay.
        Discord does **not** send this email — the hub does. Returns `202`.

        Requires the merchant to have an invite link stored (`PUT /merchants/{merchantId}/invite`);
        if none is stored, returns `409`. Idempotent re-sends are safe (the same link is re-emailed).
        Only Agents/Admins may send (FR-20).
      parameters:
        - $ref: '#/components/parameters/SpaceId'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202': { $ref: '#/components/responses/JobAccepted' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: |
            No invite-with-role link is stored for this space's merchant. Store one first via
            `PUT /merchants/{merchantId}/invite`. Body is `{code, message}`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /channels/{id}/collaborators/{userId}:
    delete:
      tags: [Collaborators]
      operationId: expelCollaborator
      summary: Expel a collaborator from a space (FR-19, async).
      description: |
        `scope=channel` (default) removes the merchant role from the member, keeping the person in
        the guild but dropping access to the merchant's channel. `scope=server` also removes them
        from the guild entirely. Audited either way. Note: because access is role-based, channel-scope
        removal drops access to the merchant's space (one space at the 1:1 cardinality).
      parameters:
        - $ref: '#/components/parameters/SpaceId'
        - $ref: '#/components/parameters/UserId'
        - name: scope
          in: query
          required: false
          description: Expulsion scope. Default `channel` (least-destructive).
          schema:
            type: string
            enum: [channel, server]
            default: channel
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202': { $ref: '#/components/responses/JobAccepted' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /collaborators/{userId}/channels:
    get:
      tags: [Collaborators]
      operationId: listCollaboratorChannels
      summary: Spaces a collaborator has access to, with merchant, role, state (FR-21).
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Spaces accessible to the collaborator.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/CollaboratorChannel' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----------------------------------------------------------------------- #
  # AGENTS (Admin only)
  # ----------------------------------------------------------------------- #
...
components:
  securitySchemes:
    ServiceApiKey:
      type: http
      scheme: bearer
      description: |
        Opaque service API key issued to the backoffice (AuthZ Layer A). Sent as
        `Authorization: Bearer <key>`. Stored hashed at rest; instantly revocable. The future POC
        frontend uses a separate session principal (not this long-lived key) behind the same
        authorization layer.
  parameters:
...
    SpaceId:
      name: id
      in: path
      required: true
      description: Space (channel) id.
      schema: { type: string, format: uuid }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Client-supplied key making a mutating request idempotent (NFR-3). A repeated key with the
        same body replays the original response; a repeated key with a different body returns 409.
      schema: { type: string, maxLength: 255 }
...
  headers:
    JobLocation:
      description: URL to poll for job completion.
      schema: { type: string, format: uri }
  responses:
    JobAccepted:
      description: |
        Accepted for asynchronous processing. The desired state is persisted; a worker will apply
        the Discord-side change. Poll the returned job for completion.
      headers:
        Location: { $ref: '#/components/headers/JobLocation' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/JobAcceptedBody' }
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid service API key.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not authorized (resolved against Postgres, NFR-13).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Idempotency-Key reused with a different body, or a conflicting state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    SpaceLifecycleState:
      type: string
      enum: [active, resolved, archived]
...
    JobStatus:
      type: string
      enum: [pending, active, completed, retrying, archived]
...
    RegisterCollaboratorRequest:
      type: object
      description: |
        Register a collaborator on a space by **name and work email** — our traceability labels,
        stored in Postgres, never a Discord lookup key. An existing collaborator may be referenced by
        `user_id` instead (e.g. one already registered on another merchant's space). A collaborator is
        a global identity, not bound to a merchant; the merchant scope comes from the target space
        (the path `id`). The same collaborator may be registered on spaces of several merchants.
      properties:
        user_id:
          type: [string, 'null']
          format: uuid
          description: Reference an existing collaborator instead of supplying name+email.
        email:
          type: [string, 'null']
          format: email
          description: The collaborator's work email; the invite link is emailed here.
        name:
          type: [string, 'null']
          description: The collaborator's name (traceability label).
      anyOf:
        - required: [user_id]
        - required: [email, name]
...
    SpaceMember:
      type: object
      required: [user_id, role]
      properties:
        user_id: { type: string, format: uuid }
        discord_user_id: { type: [string, 'null'] }
        display_name: { type: [string, 'null'] }
        merchant_id:
          type: [string, 'null']
          format: uuid
          description: The space's merchant (derived from the space, not a property of the collaborator). A collaborator is a global identity and may be a member of spaces across several merchants.
        email:
          type: [string, 'null']
          format: email
          description: The collaborator's work email (our label); the invite link is emailed here.
        role:
          type: string
          enum: [collaborator]
        invite_sent_at:
          type: [string, 'null']
          format: date-time
          description: When the merchant invite-with-role link was emailed to this collaborator (null = not yet sent).
        invited_by: { type: [string, 'null'], format: uuid }
        created_at: { type: string, format: date-time }
    CollaboratorChannel:
      type: object
      description: One space a collaborator can access. A collaborator may have several of these spanning different merchants; the merchant on each entry is the space's merchant, not the collaborator's.
      required: [space_id, role, lifecycle_state]
      properties:
        space_id: { type: string, format: uuid }
        space_name: { type: string }
        merchant_id:
          type: string
          format: uuid
          description: The space's merchant (derived from the space).
        merchant_name: { type: string }
        role:
          type: string
          enum: [collaborator]
        lifecycle_state: { $ref: '#/components/schemas/SpaceLifecycleState' }
...
    JobAcceptedBody:
      type: object
      required: [job]
      properties:
        job: { $ref: '#/components/schemas/Job' }
    Job:
      type: object
      required: [id, kind, status, created_at]
      properties:
        id: { type: string, format: uuid }
        kind:
          type: string
          description: e.g. provision_space, send_collaborator_invite, expel, project_agent_role, change_lifecycle.
        status: { $ref: '#/components/schemas/JobStatus' }
        space_id: { type: [string, 'null'], format: uuid }
        merchant_id: { type: [string, 'null'], format: uuid }
        user_id: { type: [string, 'null'], format: uuid }
        retry_count: { type: integer }
        error:
          type: [string, 'null']
          description: Last error (redacted) if retrying or archived.
        created_at: { type: string, format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code, e.g. validation_error, forbidden, not_found, idempotency_conflict.
        message: { type: string }
        details:
          type: [object, 'null']
          additionalProperties: true